Reorganise the UTTT component files into UTTT component folders.
