+++ aisoc/client/package.json
@@ -19,7 +19,8 @@
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.16.3",
-    "eslint": "^7.29.0"
+    "eslint": "^7.29.0",
+    "react-error-overlay": "6.0.9"
   },
   "scripts": {
     "start": "react-scripts start",
