+++ aisoc/client/src/competitions/uttt/components/GameCard/GameCard.scss
@@ -1,0 +1,94 @@
+@import 'common/common';
+@import 'common/fonts';
+
+.game-card {
+  background-color: #e2e8f0;
+  color: $off-black;
+  border-radius: 3px;
+
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+
+  padding: 5px 5px 5px 20px;
+  margin-bottom: 23px;
+
+  transition: transform 200ms ease-in-out, box-shadow 200ms linear;
+
+  box-shadow: 0 0 #0000, 0 0 #0000, 6px 6px #64748b;
+
+  &:hover {
+    transform: translateY(-2px);
+    box-shadow: 0 0 #0000, 0 0 #0000, 6px 6px #475569;
+  }
+
+  .mini-player {
+    display: flex;
+    align-items: center;
+
+
+    &> :not(:last-child) {
+      margin-right: 20px;
+    }
+  }
+
+  .controls {
+    button {
+      background: none;
+      color: inherit;
+      border: none;
+      padding: 0;
+      font: inherit;
+      cursor: pointer;
+      outline: inherit;
+    }
+
+    svg {
+      padding: 10px;
+      cursor: pointer;
+
+      backface-visibility: hidden;
+
+      &:hover {
+        background-color: #f1f5f9;
+        border-radius: 100%;
+        animation: control-hover-over 1.0s forwards;
+      }
+
+      &:active {
+        background-color: #cbd5e1;
+      }
+    }
+  }
+
+  .move-number {
+    color: #323232;
+    text-align: center;
+  }
+
+  .labeled-value {
+    text-align: center;
+
+    .main {
+      font-weight: bold;
+      font-size: 30px;
+      display: block;
+    }
+
+    .label {
+      color: #646464;
+    }
+  }
+
+  &.won {
+    border-left: 20px solid $win;
+  }
+
+  &.drawn {
+    border-left: 20px solid $draw;
+  }
+
+  &.lost {
+    border-left: 20px solid $loss;
+  }
+}
