+++ aisoc/client/src/competitions/uttt/components/Grid/Grid.scss
