+++ aisoc/client/src/competitions/uttt/pages/Game/Game.js
@@ -6,8 +6,8 @@ import React, { useEffect, useRef, useState } from 'react';
 import { useParams } from 'react-router';
 import { Link } from 'react-router-dom';
 import UTTTAPI from '../../api';
-import Grid from '../../components/Grid';
-import GameState from '../../services/gameReducer';
+import GameState from '../../components/Grid/GameState';
+import Grid from '../../components/Grid/Grid';
 import './Game.scss';
